added floating points

+++ fahrenheit-celsius.c
@@ -4,8 +4,8 @@
     for fahr = 0, 20, ..., 300 */
 main()
 {
-  int fahr, celsius;
-  int lower, upper, step;
+  float fahr, celsius;
+  float lower, upper, step;
 
   lower = 0;    /* lower limit of temperature scale */
   upper = 300;  /* upper limit */
@@ -15,7 +15,7 @@ main()
 
   while (fahr <= upper) {
     celsius = 5 * (fahr - 32) / 9;
-    printf("%3d\t%6d\n", fahr, celsius);
+    printf("%3.0f\t%6.1f\n", fahr, celsius);
     fahr = fahr + step;
   }
 }
